Expose option do define docker user
As we run as root, most of the times that is not who you would like to
have added to the docker group (given root already can do docker things).
One example is on Vagrant, which we would like the vagrant user to be
added on the recipe.

A caveate is that vagrant must be passed as a role before docker, so it
can define the user before anything else.